Move text feature extraction
