+++ freediscovery/engine/tests/test_near_duplicates.py
@@ -5,7 +5,7 @@ from unittest import SkipTest
 import numpy as np
 import pytest
 
-from freediscovery.text import FeatureVectorizer
+from freediscovery.engine.vectorizer import FeatureVectorizer
 from freediscovery.tests.run_suite import check_cache
