+++ freediscovery/server/tests/test_duplicate_detection.py
@@ -1,34 +1,21 @@
 # -*- coding: utf-8 -*-
 
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
-from __future__ import unicode_literals
-
-import os
 import pytest
-import json
-import itertools
 from unittest import SkipTest
-from numpy.testing import assert_equal, assert_almost_equal
 
-from .. import fd_app
-from ...utils import _silent, dict2type, sdict_keys
-from ...ingestion import DocumentIndex
-from ...exceptions import OptionalDependencyMissing
-from ...tests.run_suite import check_cache
+from freediscovery.utils import dict2type
 
 from .base import parse_res, V01, app, app_notest, get_features_cached
 
-#=============================================================================#
+# ============================================================================#
 #
 #                     Duplicates detection
 #
-#=============================================================================#
+# ============================================================================#
 
 
 @pytest.mark.parametrize('kind, options', [['simhash', {'distance': 3}],
-                                             ['i-match', {}]])
+                                           ['i-match', {}]])
 def test_api_dupdetection(app, kind, options):
 
     if kind == 'simhash':
@@ -42,9 +29,9 @@ def test_api_dupdetection(app, kind, options):
     method = V01 + "/feature-extraction/{}".format(dsid)
     data = app.get_check(method)
 
-    url = V01 + "/duplicate-detection" 
-    pars = { 'parent_id': dsid,
-             'method': kind}
+    url = V01 + "/duplicate-detection"
+    pars = {'parent_id': dsid,
+            'method': kind}
     data = app.post_check(url, json=pars)
     assert dict2type(data) == {'id': 'str'}
     mid = data['id']
