+++ freediscovery/server/tests/test_lsi.py
@@ -1,32 +1,13 @@
 # -*- coding: utf-8 -*-
 
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
-from __future__ import unicode_literals
-
-import os
-import pytest
-import json
-import itertools
-from unittest import SkipTest
-from numpy.testing import assert_equal, assert_almost_equal
-
-from .. import fd_app
-from ...utils import _silent, dict2type, sdict_keys
-from ...ingestion import DocumentIndex
-from ...exceptions import OptionalDependencyMissing
-from ...tests.run_suite import check_cache
-
 from .base import (parse_res, V01, app, app_notest,
                    get_features_cached)
 
-
-#=============================================================================#
+# ============================================================================#
 #
 #                     LSI
 #
-#=============================================================================#
+# ============================================================================#
 
 def test_api_lsi(app):
     dsid, pars, _ = get_features_cached(app)
@@ -56,4 +37,3 @@ def test_api_lsi(app):
 
     for key in data.keys():
         assert data[key] == lsi_pars[key]
-
